+++ package.json
@@ -8,13 +8,17 @@
 		"dev": "vite",
 		"build": "tsc && vite build",
 		"preview": "vite preview",
-		"macStart": "nodemon --watch './' --ext ts --exec 'node --loader ts-node/esm' server.cjs",
-		"devStart": "nodemon server.cjs"
+		"server": "nodemon --watch './' --ext ts,js --exec 'tsx' server.js",
+		"macStart": "nodemon --watch './' --ext ts --exec 'node --loader ts-node/esm' server.js",
+		"devStart": "nodemon server.cjs",
+		"dev:full": "concurrently \"npm run server\" \"npm run dev\""
 	},
 	"devDependencies": {
 		"@types/express": "^5.0.3",
+		"concurrently": "^9.2.0",
 		"dotenv": "^17.2.1",
 		"nodemon": "^3.1.10",
+		"tsx": "^4.20.3",
 		"typescript": "~5.8.3",
 		"undici-types": "^7.12.0",
 		"vite": "^7.0.4"
